refactor: consolidate HCL schema and translation logic into dedicated files
